restore gitlab-ci setting

+++ benchmark/benchmark_device_run_length_encode.parallel.hpp
@@ -58,7 +58,7 @@ inline std::string run_length_encode_config_name<rocprim::default_config>()
     return "default_config";
 }
 
-template<typename T, int MaxLength, typename Config = rocprim::default_config>
+template<typename T, size_t MaxLength, typename Config = rocprim::default_config>
 struct device_run_length_encode_benchmark : public config_autotune_interface
 {
     std::string name() const override
